beta (develop) 1.9.1 test!!!! 感谢Slime!!

+++ src/main/java/dev/rstminecraft/elytra/NoKineticDamage.java
@@ -113,6 +113,6 @@ public final class NoKineticDamage {
     }
 
     private static void setPitchSafe(Entity e) {
-        e.setPitch(safePitch((float) -89.999));
+        e.setPitch(safePitch(-89.999f));
     }
 }
